+++ src/apis/Signup/AuthCodeRequests.ts
@@ -2,7 +2,7 @@ import { apiClient } from '@/apis/apiClient';
 import { ICommonResponse } from '@/types/Common/ICommonResponse';
 import { IAuth } from '@/types/Signup/IAuth';
 
-export const createhAuth = async (): Promise<ICommonResponse<IAuth[]>> => {
+export const createhAuthCode = async (): Promise<ICommonResponse<IAuth[]>> => {
   const response = await apiClient.post('/api/Auth');
   return response.data;
 };
